Attempted link from infofragment to server, added sensor to capture fragment, updated sign in

+++ app/src/main/java/edu/cnm/deepdive/celestialbodies/model/entity/StarDetail.java
@@ -34,7 +34,7 @@ public class StarDetail {
   private Double dec;
   @SerializedName("dist")
   @Expose
-  private Integer dist;
+  private Double dist;
   @SerializedName("pmra")
   @Expose
   private Double pmra;
@@ -189,11 +189,11 @@ public class StarDetail {
     this.dec = dec;
   }
 
-  public Integer getDist() {
+  public Double getDist() {
     return dist;
   }
 
-  public void setDist(Integer dist) {
+  public void setDist(Double dist) {
     this.dist = dist;
   }
